requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-CORS==4.0.0
python-dotenv==1.0.0
face-recognition==1.3.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
Werkzeug==2.3.7
Gunicorn==21.2.0